+++ setup.sh
@@ -17,8 +17,7 @@ conda_setup() {
     conda activate rppg-toolbox || exit 1
     pip install torch==2.1.2+cu121 torchvision==0.16.2+cu121 torchaudio==2.1.2+cu121 --index-url https://download.pytorch.org/whl/cu121
     pip install -r requirements.txt || exit 1
-    cd tools/mamba || exit 1
-    python setup.py install || exit 1
+    pip install --no-build-isolation mamba-ssm causal-conv1d || exit 1
 }
 
 # Function to set up using uv
@@ -29,7 +28,7 @@ uv_setup() {
     uv pip install setuptools wheel || exit 1
     uv pip install torch==2.1.2+cu118 torchvision==0.16.2+cu118 torchaudio==2.1.2+cu118 --index-url https://download.pytorch.org/whl/cu118 || exit 1
     uv pip install -r requirements.txt --no-build-isolation || exit 1
-    cd tools/mamba && python setup.py install || exit 1
+    uv pip install --no-build-isolation mamba-ssm causal-conv1d || exit 1
     # Explicitly install PyQt5 to use interactive plotting and avoid non-interactive backends
     # See this relevant issue for more details: https://github.com/astral-sh/uv/issues/6893
     uv pip install PyQt5